current bugs: searchbar value not resetting during enter press
