+++ src/components/layout/commander/Checkbox.js
@@ -1,35 +1,26 @@
 import React, {useState, useEffect, useRef} from 'react'
 import {connect} from 'react-redux';
 
-const Checkbox = ({name, color, setColor, search: {commanders}}) => {
+const Checkbox = ({name, color, setColor}) => {
 
   const checkboxEl = useRef(null)
   const [checked, setChecked] = useState(false);
 
-  //resets checkbox to default when search form receives information
-
-  useEffect( () => {
-    checkboxEl.current.checked = false;
-  }, [commanders])
+  useEffect(() => {
+    setColor(color, checked);
+  },[checked]);
 
   const changeChecked = () => {
-    checkboxEl.current.checked = !checkboxEl.current.checked;
-    if(checkboxEl.current.checked){
-      setChecked(true);
-    }
-    setColor(color, checkboxEl.current.checked);
+    setChecked(!checked);
   }
 
   return (
     <div className={`flex-container-row checkbox checkbox-${color}`} onClick={changeChecked}>
       <label htmlFor={color}>{name}</label>
-      <input ref={checkboxEl} type="checkbox" name={color} checked={false} onChange={() => {}}/>
+      <input ref={checkboxEl} type="checkbox" name={color} checked={checked}/>
       <span className='checkmark'></span>
     </div>
   )
 }
 
-const mapStateToProps = state => ({
-  search: state.search
-})
-export default connect(mapStateToProps)(Checkbox)
+export default connect()(Checkbox)
